+++ system/lib/actor_system.dart
@@ -4,4 +4,5 @@ export 'src/system/context.dart'
     show BaseContext, ActorContext, ActorSystem, MissingHostHandling, PathMatcher, patternMatcher;
 export 'src/system/exceptions.dart' show SkipMessage, MessageNotDelivered, ActorStopped, MailboxFull, NoFactoryFound;
 export 'src/system/messages.dart' show initMsg, shutdownMsg;
+export 'src/system/metrics.dart' show ActorSystemMetrics;
 export 'src/system/ref.dart' show ActorRef;